refactor: split restart command module
- extract /restart command flow from chat-commands

- move restart status persistence and worker script into dedicated file

- keep command routing in chat-commands with minimal behavior change
